theme.js refactoring
